[refactor] magi-human: defer audio VAE to main's stable-audio infra
Drop the magi-introduced 'encoder-style' Stable Audio VAE wrapper now
that main has merged a first-class Oobleck VAE port plus a shared
SAAudioVAEModel pipeline-glue lazy-loader (#1260). MagiHuman now
reuses that infrastructure instead of carrying duplicates:

  - audio VAE config:  fastvideo.configs.models.vaes.OobleckVAEConfig
  - audio VAE wrapper: fastvideo.models.vaes.sa_audio.SAAudioVAEModel

Removes:
  - fastvideo/configs/models/encoders/sa_audio.py (SAAudioVAEConfig)
  - fastvideo/models/encoders/sa_audio.py        (SAAudioVAEModel dup)

Updates:
  - magi-human pipeline + config to construct OobleckVAEConfig and set
    pretrained_path instead of arch_config.sa_audio_model_path.
  - encoders/__init__.py drops the SAAudioVAEConfig export.
  - parity test moves from tests/local_tests/encoders/ to .../vaes/
    (matches main's classification) and switches to the shared wrapper;
    explicit pretrained_dtype='float32' override keeps the fp32 ref
    parity check intact (default is fp16 to match official
    stable_audio_tools).

audio_decoding.py docstring still mentions 'sa_audio_vae_model' — main's
wrapper exposes that name as a back-compat alias for oobleck_vae, so
the existing comment is still accurate.
